Refactor yet again customizing the different interfaces
