+++ src/main/java/org/alliancegenome/curation_api/interfaces/base/crud/BaseDeleteCurieControllerInterface.java
@@ -1,4 +1,4 @@
-package org.alliancegenome.curation_api.interfaces.base;
+package org.alliancegenome.curation_api.interfaces.base.crud;
 
 import org.alliancegenome.curation_api.model.entities.base.AuditedObject;
 import org.alliancegenome.curation_api.response.ObjectResponse;
