+++ src/main/java/org/alliancegenome/curation_api/interfaces/base/crud/BaseReadIdControllerInterface.java
@@ -1,8 +1,9 @@
-package org.alliancegenome.curation_api.interfaces.base;
+package org.alliancegenome.curation_api.interfaces.base.crud;
 
 import org.alliancegenome.curation_api.model.entities.base.AuditedObject;
 import org.alliancegenome.curation_api.response.ObjectResponse;
 import org.alliancegenome.curation_api.view.View;
+import org.apache.commons.lang3.ObjectUtils.Null;
 import org.eclipse.microprofile.openapi.annotations.media.Content;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
@@ -10,9 +11,8 @@ import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;
 
 import com.fasterxml.jackson.annotation.JsonView;
 
-import jakarta.validation.constraints.Null;
 import jakarta.ws.rs.Consumes;
-import jakarta.ws.rs.DELETE;
+import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
@@ -20,19 +20,20 @@ import jakarta.ws.rs.core.MediaType;
 
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-public interface BaseDeleteIdentifierControllerInterface<E extends AuditedObject> {
+public interface BaseReadIdControllerInterface<E extends AuditedObject> {
 
-	@DELETE
-	@Path("/{identifierString}")
+	@GET
+	@Path("/{id}")
 	@JsonView(View.FieldsOnly.class)
 	@APIResponses(
 		@APIResponse(
-			description = "Delete the Entity by Identifier String",
+			description = "Get Entity by Id",
 			content = @Content(
 				mediaType = "application/json",
 				schema = @Schema(implementation = Null.class)
 			)
 		)
 	)
-	public ObjectResponse<E> deleteByIdentifier(@PathParam("identifierString") String identifierString);
+	public ObjectResponse<E> getById(@PathParam("id") Long id);
+	
 }
