+++ src/main/java/org/alliancegenome/curation_api/interfaces/crud/BiologicalEntityCrudInterface.java
@@ -1,6 +1,6 @@
 package org.alliancegenome.curation_api.interfaces.crud;
 
-import org.alliancegenome.curation_api.interfaces.base.SubmittedObjectCrudInterface;
+import org.alliancegenome.curation_api.interfaces.base.BaseSubmittedObjectCrudInterface;
 import org.alliancegenome.curation_api.model.entities.BiologicalEntity;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 
@@ -13,6 +13,6 @@ import jakarta.ws.rs.core.MediaType;
 @Tag(name = "CRUD - Biological Entities")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-public interface BiologicalEntityCrudInterface extends SubmittedObjectCrudInterface<BiologicalEntity> {
+public interface BiologicalEntityCrudInterface extends BaseSubmittedObjectCrudInterface<BiologicalEntity> {
 
 }
