+++ src/main/java/org/alliancegenome/curation_api/interfaces/crud/ConstructCrudInterface.java
@@ -3,7 +3,7 @@ package org.alliancegenome.curation_api.interfaces.crud;
 import java.util.List;
 
 import org.alliancegenome.curation_api.interfaces.base.BaseUpsertControllerInterface;
-import org.alliancegenome.curation_api.interfaces.base.SubmittedObjectCrudInterface;
+import org.alliancegenome.curation_api.interfaces.base.BaseSubmittedObjectCrudInterface;
 import org.alliancegenome.curation_api.model.entities.Construct;
 import org.alliancegenome.curation_api.model.ingest.dto.ConstructDTO;
 import org.alliancegenome.curation_api.response.APIResponse;
@@ -26,7 +26,7 @@ import jakarta.ws.rs.core.MediaType;
 @Tag(name = "CRUD - Constructs")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-public interface ConstructCrudInterface extends SubmittedObjectCrudInterface<Construct>, BaseUpsertControllerInterface<Construct, ConstructDTO> {
+public interface ConstructCrudInterface extends BaseSubmittedObjectCrudInterface<Construct>, BaseUpsertControllerInterface<Construct, ConstructDTO> {
 
 	@Override
 	@GET
