+++ src/main/java/org/alliancegenome/curation_api/services/base/SubmittedObjectCrudService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import org.alliancegenome.curation_api.dao.base.BaseEntityDAO;
 import org.alliancegenome.curation_api.enums.BackendBulkDataProvider;
 import org.alliancegenome.curation_api.exceptions.ObjectUpdateException;
-import org.alliancegenome.curation_api.interfaces.base.BaseDeleteIdentifierControllerInterface;
-import org.alliancegenome.curation_api.interfaces.base.BaseReadIdControllerInterface;
 import org.alliancegenome.curation_api.interfaces.base.BaseUpsertControllerInterface;
+import org.alliancegenome.curation_api.interfaces.base.crud.BaseDeleteIdentifierControllerInterface;
+import org.alliancegenome.curation_api.interfaces.base.crud.BaseReadIdControllerInterface;
 import org.alliancegenome.curation_api.model.entities.base.SubmittedObject;
 import org.alliancegenome.curation_api.model.ingest.dto.base.BaseDTO;
 import org.alliancegenome.curation_api.response.ObjectResponse;
